+++ actionpack/test/controller/send_file_test.rb
@@ -210,7 +210,7 @@ class SendFileTest < ActionController::TestCase
   def test_send_file_instrumentation
     @controller.options = { disposition: :inline }
 
-    assert_notification("send_file.action_controller", path: __FILE__, disposition: :inline) do
+    assert_notification("send_file.action_controller", { path: __FILE__, disposition: :inline }) do
       process("file")
     end
   end
@@ -218,7 +218,7 @@ class SendFileTest < ActionController::TestCase
   def test_send_data_instrumentation
     @controller.options = { content_type: "application/x-ruby" }
 
-    assert_notification("send_data.action_controller", content_type: "application/x-ruby") do
+    assert_notification("send_data.action_controller", { content_type: "application/x-ruby" }) do
       process("data")
     end
   end